+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 * Add compatibility with Sprockets 4, see [Sprockets documentation](https://github.com/rails/sprockets/blob/master/UPGRADING.md) on how to upgrade your Rails app ([PR #2691](https://github.com/alphagov/govuk_publishing_components/pull/2691))
 * Use a data attribute to specify RUM (real user monitoring) script location ([PR #2682](https://github.com/alphagov/govuk_publishing_components/pull/2682))
+* Add HMRC email to attachment accessible format request pilot([PR #2695](https://github.com/alphagov/govuk_publishing_components/pull/2695))
 
 ## 28.9.2